# spec/openapi.yaml
openapi: 3.0.3
info:
  title: SIRISA UNISM SSO API
  version: 2.8.1
  description: |
    Dokumentasi **read-only** API integrasi client UNISM (SISKA, Presensi, SANTI, dll.).

    **Penting untuk client apps:**
    - Spec ini **hanya dokumentasi** — tidak mengubah endpoint, request, atau response yang sudah berjalan.
    - Dua endpoint verify-token **tetap aktif** (backward compatible).
    - Scope `access-user` tetap memberi full access seperti sebelum P1.
    - Client lama **tidak wajib** migrasi hanya karena adanya dokumentasi ini.

    OAuth2 token diperoleh via flow Passport standar (`/oauth/authorize`, `/oauth/token`).
    Semua endpoint terproteksi memakai header `Authorization: Bearer <access_token>`.

  contact:
    name: ICT UNISM
  license:
    name: Proprietary

servers:
  - url: '{baseUrl}/api'
    description: API base (prefix Laravel `/api`)
    variables:
      baseUrl:
        default: https://sirisa.unism.ac.id
        description: URL aplikasi SSO (tanpa trailing slash)

tags:
  - name: Token Verification
    description: Verifikasi bearer token — **tanpa scope middleware** (rate limit 30/menit per IP)
  - name: User Read
    description: Membutuhkan scope `read-user` **atau** `access-user`
  - name: User Write
    description: Membutuhkan scope `write-user` **atau** `access-user`

paths:
  /verify-token:
    get:
      tags: [Token Verification]
      summary: Verifikasi token (response minimal)
      description: |
        Endpoint canonical untuk cek validitas token OAuth tanpa data user/role.
        Dipakai client apps untuk validasi ringan.
      operationId: verifyTokenMinimal
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenMinimalSuccess'
              example:
                valid: true
                token_id: 123
                user_id: 45
                scopes: ['access-user']
        '401':
          description: Token tidak ada / tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenError'
              examples:
                noToken:
                  value:
                    valid: false
                    error: No token provided
                invalid:
                  value:
                    valid: false
                    error: The resource owner or authorization server denied the request.

  /authorize/verify-token:
    get:
      tags: [Token Verification]
      summary: Verifikasi token (response lengkap)
      description: |
        Response lengkap dengan username, role_id, dan client_id.
        **Tetap dipertahankan** untuk backward compatibility dengan client apps existing.
      operationId: verifyTokenFull
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenFullSuccess'
              example:
                status: true
                message: Token is valid
                data:
                  token_id: 123
                  username: 1234567890
                  role: 63
                  client_id: 1
        '401':
          description: Token tidak valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifyTokenError'

  /user:
    get:
      tags: [User Read]
      summary: Profil user pemilik token
      description: |
        Mengembalikan data user yang terautentikasi beserta relasi `oauth_client_users`.
        Dipakai client apps (mis. Presensi `GET /sso/profile`) untuk refresh avatar/profil.
      operationId: getAuthenticatedUser
      security:
        - bearerAuth: [read-user, access-user]
      responses:
        '200':
          description: Data user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
    post:
      tags: [User Write]
      summary: Buat user baru
      operationId: createUser
      security:
        - bearerAuth: [write-user, access-user]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreateRequest'
      responses:
        '201':
          description: User dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
        '422':
          $ref: '#/components/responses/ValidationError'

  /username:
    get:
      tags: [User Read]
      summary: Cari user by username
      description: |
        Query parameter `username` wajib. Dipakai SISKA untuk cek keberadaan user & nomor HP.
      operationId: getUserByUsername
      security:
        - bearerAuth: [read-user, access-user]
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
          example: '1234567890'
      responses:
        '200':
          description: User ditemukan
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/User'
        '404':
          description: User tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: User not found
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'

  /user/{username}/roles:
    get:
      tags: [User Read]
      summary: Daftar role OAuth user
      operationId: getUserRoles
      security:
        - bearerAuth: [read-user, access-user]
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      responses:
        '200':
          description: Daftar oauth client user + role
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/OauthClientUser'
        '404':
          description: User tidak ditemukan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'

  /oauthClientUsers:
    post:
      tags: [User Write]
      summary: Assign role OAuth ke user
      operationId: createOauthClientUser
      security:
        - bearerAuth: [write-user, access-user]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, oauth_client_role_id]
              properties:
                user_id:
                  type: integer
                oauth_client_role_id:
                  type: integer
      responses:
        '201':
          description: Relasi dibuat
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OauthClientUser'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
        '422':
          $ref: '#/components/responses/ValidationError'

  /user/{username}/{newUsername}:
    put:
      tags: [User Write]
      summary: Update data user
      description: Update profil user tanpa mengubah role (oauth_client_role_id hanya referensi validasi).
      operationId: updateUser
      security:
        - bearerAuth: [write-user, access-user]
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
        - name: newUsername
          in: path
          required: true
          schema:
            type: string
          description: Username baru (bisa sama dengan username lama)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdateRequest'
      responses:
        '200':
          description: Berhasil diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: User updated successfully
        '404':
          description: User tidak ditemukan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'
        '422':
          $ref: '#/components/responses/ValidationError'

  /user/actived/{username}:
    post:
      tags: [User Write]
      summary: Update status aktif user
      operationId: updateUserActive
      security:
        - bearerAuth: [write-user, access-user]
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [is_active]
              properties:
                is_active:
                  type: boolean
      responses:
        '200':
          description: Status diupdate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: User Active updated successfully
        '404':
          description: User tidak ditemukan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'

  /user/{username}/cleanup-duplicates:
    post:
      tags: [User Write]
      summary: Hapus duplikasi role user
      operationId: cleanupDuplicateRoles
      security:
        - bearerAuth: [write-user, access-user]
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      responses:
        '200':
          description: Cleanup selesai
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  duplicatesRemoved:
                    type: integer
        '404':
          description: User tidak ditemukan
        '500':
          description: Cleanup gagal
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'

  /user/{username}:
    delete:
      tags: [User Write]
      summary: Hapus role user / user
      description: Menghapus relasi oauth_client_users; jika tidak ada role tersisa, user ikut dihapus.
      operationId: deleteUserRole
      security:
        - bearerAuth: [write-user, access-user]
      parameters:
        - $ref: '#/components/parameters/UsernamePath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [oauth_client_role_id]
              properties:
                oauth_client_role_id:
                  type: integer
      responses:
        '200':
          description: Role dihapus
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
              example:
                message: Role user berhasil dihapus
        '404':
          description: User tidak ditemukan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/ForbiddenScope'

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 Access Token
      description: |
        Passport OAuth2 access token.

        **Scopes (Passport):**
        | Scope | Akses |
        |-------|-------|
        | `verify-token` | Dokumentasi — verifikasi via bearer (endpoint verify-token tidak pakai scope middleware) |
        | `read-user` | GET user, username, roles |
        | `write-user` | POST/PUT/DELETE user & assignment |
        | `access-user` | **Full access** — backward compatible client lama |

  parameters:
    UsernamePath:
      name: username
      in: path
      required: true
      schema:
        type: string

  responses:
    Unauthorized:
      description: Tidak terautentikasi
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
    ForbiddenScope:
      description: Token valid tetapi scope tidak cukup
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/MessageResponse'
          example:
            message: Invalid scope(s) provided.
    ValidationError:
      description: Validasi gagal
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
              errors:
                type: object
                additionalProperties:
                  type: array
                  items:
                    type: string

  schemas:
    VerifyTokenMinimalSuccess:
      type: object
      required: [valid, token_id, user_id, scopes]
      properties:
        valid:
          type: boolean
          example: true
        token_id:
          type: integer
        user_id:
          type: integer
        scopes:
          type: array
          items:
            type: string

    VerifyTokenFullSuccess:
      type: object
      required: [status, message, data]
      properties:
        status:
          type: boolean
        message:
          type: string
        data:
          type: object
          properties:
            token_id:
              type: integer
            username:
              type: string
            role:
              type: integer
              description: oauth_client_role.role_id
            client_id:
              type: integer

    VerifyTokenError:
      type: object
      properties:
        valid:
          type: boolean
          example: false
        error:
          type: string

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    User:
      type: object
      description: Model user — field tambahan dapat ada tanpa breaking change
      properties:
        id:
          type: integer
        name:
          type: string
        username:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        is_client:
          type: boolean
        is_active:
          type: boolean
        prodi:
          type: string
          nullable: true
        foto:
          type: string
          nullable: true
        email_verified_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    AuthenticatedUser:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            avatar_url:
              type: string
              description: URL foto profil (S3 pre-signed atau proxy)
            oauth_client_users:
              type: array
              items:
                $ref: '#/components/schemas/OauthClientUser'

    OauthClientUser:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        oauth_client_role_id:
          type: integer
        oauth_client_role:
          type: object
          description: Relasi nested (client + role)

    UserCreateRequest:
      type: object
      required: [name, username, password, is_client, is_active]
      properties:
        name:
          type: string
        username:
          type: string
        password:
          type: string
          minLength: 8
        is_client:
          type: boolean
        is_active:
          type: boolean
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        prodi:
          type: string
          nullable: true

    UserUpdateRequest:
      type: object
      required: [name, oauth_client_role_id]
      properties:
        name:
          type: string
        username:
          type: string
          description: Wajib jika newUsername berbeda dari username lama
        phone:
          type: string
          nullable: true
        email:
          type: string
          format: email
          nullable: true
        prodi:
          type: string
          nullable: true
        oauth_client_role_id:
          type: integer